Split out main server to avoid creating logfile
